+++ CHANGELOG.md
@@ -16,6 +16,9 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :boom: Breaking Change
 
 * Removed params merging when navigation to the route with the same name that current route in `b-router`
+
+#### :bug: Bug Fix
+
 * Fixed alias priority, when resolve path parameters in `core/router`
 
 ## v3.52.0 (2023-07-19)