+++ CHANGELOG.md
@@ -11,11 +11,15 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v4.??.?? (2023-??-??)
+## v4.0.0-beta.10 (2023-07-27)
 
 #### :boom: Breaking Change
 
 * Merging of parameters when navigating to a route with the same name as the current one has been removed `bRouter`
+
+#### :rocket: New Feature
+
+* Added storybook support
 
 #### :bug: Bug Fix